backend/requirements.txt
fastapi
uvicorn[standard]
ultralytics
Pillow
numpy
python-multipart
torch
torchvision
